requirements.txt
requests
python-dotenv
atproto
